+++ frontend/public/components/app.jsx
@@ -14,7 +14,8 @@ import { useTranslation } from 'react-i18next';
 import { coFetchJSON, appInternalFetch } from '../co-fetch';
 import { detectFeatures, setFlag } from '../actions/features';
 import AppContents from './app-contents';
-import { getBrandingDetails, Masthead } from './masthead';
+import { Masthead } from './masthead';
+import { getBrandingDetails } from './masthead-branding';
 import { ConsoleNotifier } from './console-notifier';
 import { ConnectedNotificationDrawer } from './notification-drawer';
 import { Navigation } from '@console/app/src/components/nav';
